Move get_unit_registers() into header file.
