+++ pytket/binders/circuit/Circuit/main.cpp
@@ -24,6 +24,7 @@
 #include "UnitRegister.hpp"
 #include "binder_json.hpp"
 #include "boost/graph/iteration_macros.hpp"
+#include "circuit_registers.hpp"
 #include "deleted_hash.hpp"
 #include "py_operators.hpp"
 #include "tket/Circuit/Boxes.hpp"
@@ -56,43 +57,6 @@ UnitID to_cpp_unitid(const PyUnitID &py_unitid) {
     return get<Qubit>(py_unitid);
   }
   return get<Bit>(py_unitid);
-}
-
-template <typename T>
-std::vector<T> get_unit_registers(Circuit &circ) {
-  static_assert(
-      std::is_same<T, QubitRegister>::value ||
-          std::is_same<T, BitRegister>::value,
-      "T must be either QubitRegister or BitRegister");
-  using T2 = typename std::conditional<
-      std::is_same<T, QubitRegister>::value, Qubit, Bit>::type;
-  std::vector<T2> unitids;
-  if constexpr (std::is_same<T, QubitRegister>::value) {
-    unitids = circ.all_qubits();
-  } else if constexpr (std::is_same<T, BitRegister>::value) {
-    unitids = circ.all_bits();
-  }
-  // map from register name to unsigned indices
-  std::map<std::string, std::set<unsigned>> unit_map;
-  std::vector<T> regs;
-  for (const T2 &unitid : unitids) {
-    // UnitRegisters only describe registers with 1-d indices
-    if (unitid.reg_dim() > 1) continue;
-    auto it = unit_map.find(unitid.reg_name());
-    if (it == unit_map.end()) {
-      unit_map.insert({unitid.reg_name(), {unitid.index()[0]}});
-    } else {
-      it->second.insert(unitid.index()[0]);
-    }
-  }
-  regs.reserve(unit_map.size());
-  for (auto const &it : unit_map) {
-    // only return registers that are indexed consecutively from zero
-    if (*it.second.rbegin() == it.second.size() - 1) {
-      regs.emplace_back(it.first, it.second.size());
-    }
-  }
-  return regs;
 }
 
 void init_circuit_add_op(py::class_<Circuit, std::shared_ptr<Circuit>> &c);
